Corrections

+++ Part1/CONTROLINFO.h
@@ -7,14 +7,14 @@
 
 typedef struct
 {
-	size_t fileId;	//File identifier
-	size_t filePosition;
-	size_t nCharacters; //Number of characteres 
-	size_t nConsoants; //Number of consoants
-	size_t largestWord; //Number of characters of the largest word
-	size_t numBytes;
-	size_t wordCount; //Number of words
-	size_t nWordsBySize[MAX_SIZE_WORD]; //Number of words for each word size
+	int fileId;	//File identifier
+	int filePosition;
+	int nCharacters; //Number of characteres 
+	int nConsoants; //Number of consoants
+	int largestWord; //Number of characters of the largest word
+	int numBytes;
+	int wordCount; //Number of words
+	int nWordsBySize[MAX_SIZE_WORD]; //Number of words for each word size
 	int nConsoantsByWordLength[MAX_SIZE_WORD][MAX_SIZE_WORD]; //Number of consoants for each word size 
 }CONTROLINFO;
